added "sdcard_format" to various apps

+++ apps/processing/midibox_cv_v2/src/components/MbCvMidiVoice.h
@@ -59,7 +59,7 @@ public:
     u8 midivoiceAftertouch;
     u8 midivoiceCCNumber;
     u8 midivoiceCCValue;
-    u8 midivoiceNRPNValue;
+    u16 midivoiceNRPNValue;
 
     // stacks
     notestack_t midivoiceNotestack;
